removing unused conda env ymls and adding conda to default profile for parser usage and adding input registry table
