+++ latch_metadata/__init__.py
@@ -2,11 +2,16 @@
 from latch.types.metadata import (
     NextflowMetadata,
     LatchAuthor,
-    NextflowRuntimeResources
+    NextflowRuntimeResources, LatchParameter, LatchMetadata
 )
 from latch.types.directory import LatchDir
+from latch.registry.table import Table
+from latch.resources.tasks import small_task
 
 from .parameters import generated_parameters
+
+# is the LatchMetadata compatible with the NextflowMetadata?
+# do we need to swap to LatchMetadata if we want to have the table id input?
 
 NextflowMetadata(
     display_name='TheiaProk_Illumina_PE',
@@ -21,3 +26,4 @@ NextflowMetadata(
     ),
     log_dir=LatchDir("latch:///theiaprok_pe_log"),
 )
+
