+++ modules/nf-core/bakta/bakta/main.nf
@@ -2,7 +2,6 @@ process BAKTA_BAKTA {
     tag "$meta.id"
     label 'process_medium'
 
-    conda "${moduleDir}/environment.yml"
     container "${ workflow.containerEngine == 'singularity' && !task.ext.singularity_pull_docker_container ?
         'https://depot.galaxyproject.org/singularity/bakta:1.10.4--pyhdfd78af_0' :
         'biocontainers/bakta:1.10.4--pyhdfd78af_0' }"
