+++ modules/nf-core/prokka/main.nf
@@ -2,7 +2,6 @@ process PROKKA {
     tag "${meta.id}"
     label 'process_low'
 
-    conda "${moduleDir}/environment.yml"
     container "${ workflow.containerEngine == 'singularity' && !task.ext.singularity_pull_docker_container ?
         'https://community-cr-prod.seqera.io/docker/registry/v2/blobs/sha256/3a/3af46b047c8fe84112adeaecf300878217c629b97f111f923ecf327656ddd141/data' :
         'community.wave.seqera.io/library/prokka_openjdk:10546cadeef11472' }"
